Complete subpage Sobre Nosotros
Created layout for sobre nosotros subpage.

It shows a vertical card design for mobile and tablet resolutions, box-shadow on the bottom left of the card.

Header is floating between navbar and content.

For desktop version, the information is displayed in horizontal display, reversing each two cards, box shadow on images.